+++ R/preprocessing.R
@@ -654,13 +654,13 @@ sn_standardize_gene_symbols <- function(
 #' if (requireNamespace("Seurat", quietly = TRUE)) {
 #'   data("pbmc_small", package = "Shennong")
 #'   pbmc_filtered <- sn_filter_genes(
-#'     pbmc,
+#'     pbmc_small,
 #'     min_cells = 5,
 #'     plot = FALSE,
 #'     filter = TRUE
 #'   )
 #'   pbmc_coding <- sn_filter_genes(
-#'     pbmc,
+#'     pbmc_small,
 #'     min_cells = 1,
 #'     plot = FALSE,
 #'     filter = TRUE,